Add first draft of woodpecker-ci configuration
I am able to do the initial configuration of Gitea once it is up and
running. Then I can manually set up an OAuth application for the
woodpecker server. Finally, I can manually add a repository with a
simple .woodpecker.yml that echoes hello world and the pipeline will
show up in woodpecker.