Enable privileged access for woodpecker agent
I wasn't able to get the debug probe working without this. Only exposing
/dev/ttyACM0 does not work for some reason, I tried with various Ubuntu
and Debian base images, to no avail. I hate the idea of privileged
access but there isn't much I can do at this point. I will have to find
a way to fix this in the future.